admin can not be deleted from system

+++ service/src/main/java/cz/fi/muni/PA165/service/UserServiceImpl.java
@@ -65,6 +65,8 @@ public class UserServiceImpl implements UserService{
     @Override
     public void deleteUser(Long id) {
         User user = findById(id);
+        if (user.isAdmin())
+            throw new DnDServiceException("Admin can not be deleted.");
         userDao.delete(user);
     }
